added go docs, to body,header,cookie,query,param and deleted constants file
